+++ packages/components/organisms/f-registration/stories/Registration.stories.js
@@ -1,11 +1,16 @@
 import { withKnobs, select, text, boolean } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import Registration from '../src/components/Registration.vue';
+import RegistrationMock from '../src/mocks/registrationMock';
 
 export default {
     title: 'Components/Organisms',
     decorators: [withKnobs, withA11y]
 };
+
+const createAccountUrl = '/account/register';
+RegistrationMock.setupEmailInUse(createAccountUrl)
+RegistrationMock.passThroughAny();
 
 export const RegistrationComponent = () => ({
     components: { Registration },
@@ -17,7 +22,7 @@ export const RegistrationComponent = () => ({
             default: text('Title', 'Create Account')
         },
         createAccountUrl: {
-            default: text('Create Account URL', '/account/register')
+            default: text('Create Account URL', createAccountUrl)
         },
         buttonText: {
             default: text('Button Text', 'Create Account')
@@ -26,7 +31,7 @@ export const RegistrationComponent = () => ({
             default: boolean('Show Login Link', true)
         },
         loginUrl: {
-            default: text('Login URL', '/account/register')
+            default: text('Login URL', createAccountUrl)
         }
     },
     parameters: {